refactored some code regarding children rendering
