+++ packages/react-live-props/src/Renderers/index.js
@@ -1,3 +1,4 @@
 export { default as FieldRenderer } from './Field'
 export { default as PrimitiveArrayRenderer } from './PrimitiveArray'
 export { default as PropertyRenderer, ObjectArrayRenderer, ObjectRenderer, AddHtmlAttributeRenderer } from './Property'
+export { ChildrenObjectRenderer, ChildrenArrayRenderer } from './Children'
